Mejores colisiones

+++ includes/Mesada.h
@@ -6,9 +6,8 @@ class Mesada : public Espacio {
 private:
     // TODO: Plato sobre mesada
 public:
-    Mesada(sf::Vector2i pos, int rot, sf::Vector2i size);
+    Mesada(sf::Vector2i pos, int rot, sf::Vector2f size);
     ~Mesada();
-
     bool ponerOSacarPlato();
 };
