+++ includes/Suelo.h
@@ -5,7 +5,7 @@
 
 class Suelo : public Espacio {
 public:
-    Suelo(sf::Vector2i pos, int rot, sf::Vector2i size);
+    Suelo(sf::Vector2i pos, int rot, sf::Vector2f size);
 };
 
 #endif // SUELO_H
